fixed code climate issues
